refactor: inline wide post layout
