+++ src/components/organisms/PostInput/PostInput.tsx
@@ -20,8 +20,7 @@ import { MentionPopover } from '@/molecules/MentionPopover/MentionPopover';
 import { PostInputAttachments } from '@/molecules/PostInputAttachments/PostInputAttachments';
 import { PostPreviewCard } from '@/molecules/PostPreviewCard/PostPreviewCard';
 import { useToast } from '@/molecules/Toaster/use-toast';
-import { usePostMainLayout } from '@/organisms/PostMain/PostMainLayout';
-import { WIDE_POST_BODY_TEXT_CLASS } from '@/organisms/PostMain/WidePostLayout';
+import { usePostMainLayout, WIDE_POST_BODY_TEXT_CLASS } from '@/organisms/PostMain/PostMainLayout';
 import { PostHeader } from '../PostHeader/PostHeader';
 import { PostInputExpandableSection } from '../PostInputExpandableSection/PostInputExpandableSection';
 import { POST_INPUT_VARIANT } from './PostInput.constants';
