+++ src/components/organisms/QuickReply/QuickReply.tsx
@@ -18,8 +18,7 @@ import { canSubmitPost, cn, getCharacterCount } from '@/libs/utils/utils';
 import { MentionPopover } from '@/molecules/MentionPopover/MentionPopover';
 import { PostInputAttachments } from '@/molecules/PostInputAttachments/PostInputAttachments';
 import { POST_INPUT_VARIANT } from '@/organisms/PostInput/PostInput.constants';
-import { usePostMainLayout } from '@/organisms/PostMain/PostMainLayout';
-import { WIDE_POST_BODY_TEXT_CLASS } from '@/organisms/PostMain/WidePostLayout';
+import { usePostMainLayout, WIDE_POST_BODY_TEXT_CLASS } from '@/organisms/PostMain/PostMainLayout';
 import { AvatarWithFallback } from '../AvatarWithFallback/AvatarWithFallback';
 import { PostInputExpandableSection } from '../PostInputExpandableSection/PostInputExpandableSection';
 import { QUICK_REPLY_CONNECTOR_SPACER_HEIGHT } from './QuickReply.constants';
